pyproject.toml
[project]
name = "coco-tools"
version = "0.19"
authors = [
  {name = "Jamie Cho", email = "jamieleecho@gmail.com"},
]
description = "TRS-80 Color Computer Tools"
readme = "README.md"
classifiers=[
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Build Tools",
  "Programming Language :: Python :: 3",
]
license = "GPL-2.0"
keywords = [
  "coco",
  "image",
  "conversion",
  "trs-80",
  "tandy",
]
requires-python = ">=3.9.19"
dependencies = [
  "parsimonious>=0.10.0",
  "Pillow>=7.0.0",
  "pypng>=0.0.18",
  "pydantic>=2.0.0",
  "pydantic-yaml>=1.3.0",
]

[project.scripts  ]
  decb-to-b09 = "coco.decb_to_b09:main"
  cm3toppm = "coco.cm3toppm:main"
  hrstoppm = "coco.hrstoppm:main"
  maxtoppm = "coco.maxtoppm:main"
  mgetoppm = "coco.mgetoppm:main"
  mge_viewer2 = "coco.mge_viewer2:main"
  pixtopgm = "coco.pixtopgm:main"
  rattoppm = "coco.rattoppm:main"
  veftopng = "coco.veftopng:main"

[project.urls]
Repository = "https://github.com/jamieleecho/coco-tools"
Issues = "https://github.com/jamieleecho/coco-tools/issues"

[dependency-groups]
dev = [
  "imagehash",
  "pytest",
  "pytest-cov",
  "pre-commit",
  "ruff",
  "tk",
  "types-setuptools",
]

[build-system]
requires = [
  "setuptools",
  "scipy",
  "wheel",
]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests.coco_tests"]

[tool.setuptools.package-dir]
"" = "."

[tool.setuptools.package-data]
"coco.resources" = ["*.b09"]

[tool.pytest.ini_options]
rootdir = "."
pythonpath = [
  ".",
]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "quick: marks tests as fast",
]

[tool.ruff]
exclude = [
  "docs/",
]
